requirements.txt
# Core image processing
opencv-python==4.8.1.78
pillow==10.0.1
numpy==1.24.3
matplotlib==3.7.2

# Machine learning and AI
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
diffusers==0.21.4

# VLM APIs
openai==1.3.5
anthropic==0.7.7

# Image similarity metrics
lpips==0.1.4
scikit-image==0.21.0

# Text processing
nltk==3.8.1
textstat==0.7.3

# Data handling
pandas==2.0.3
pyyaml==6.0.1

# Progress bars and utilities
tqdm==4.66.1
requests==2.31.0

# Development and testing
jupyter==1.0.0
pytest==7.4.2

# Optional: for advanced text analysis
spacy==3.6.1